+++ R/FuzzyBayesRule.R
@@ -5,13 +5,13 @@
 #'
 #' @param train matrix or data frame of training set cases.
 #' @param cl factor of true classifications of training set
-#' @param cores  how many cores of the computer do you want to use (default = 2)
+#' @param cores  how many cores of the computer do you want to use to use for prediction (default = 2)
 #' @param fuzzy boolean variable to use the membership function
 #'
 #' @return A vector of classifications
 #'
 #' @references
-#' \insertRef{roneide2006fuzzy}{FuzzyClass}
+#' \insertRef{de2006fuzzy}{FuzzyClass}
 #'
 #' @examples
 #'